requirements.txt
watchdog==0.9.0
requests==2.21.0
python-dotenv==0.10.1